refactor(DownloadPriceService):downloadPrice()

+++ database/migrations/2022_01_04_151444_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name', 255)->nullable(false);
-            $table->integer('price')->default(0);
+            $table->float('price')->default(0);
             $table->timestamps();
         });
     }
